Boutons retry (principalement)

+++ ConsoleApplication2/Board.cpp
@@ -60,7 +60,7 @@ void Board::update(sf::Clock *clock) {
 
 float Board::getScore()
 {
-	float score = 0;
+	int score = 0;
 	score += snake.getLength();
 	//score depends on the time spent as well ? 
 	return score;
